correct documentation of `F` parameter to nlsolve
As Kevin pointed out, the termination condition of nlsolve solve is
f(x) = 0, not f(x) = F as I had assumed.

More broadly, the examples should perhaps be updated to clarify that
the initial values in such scratch buffers are just placeholders.

(This is a place where C++ seems to do things well - often, the
convention is to pass immutable arguments as `const` reference, and
mutable arguments as pointers to non-`const`. This makes it clear, on
the caller and callee side, as to which particular arguments to a
function are (im)mutable, instead of Julia's current `!` function
suffix that only indicates whether _any_ arguments are mutable, but
not which ones. At least from a cursory look, it seems Julia doesn't
yet have such a feature.)